requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langgraph>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
asyncio
fastapi
uvicorn
streamlit
faiss-cpu
sentence-transformers
requests
